+++ Source/LandscapeHeightmapTracker/Private/LandscapeHeightmapTrackerEdMode.h
@@ -7,6 +7,7 @@ class FLandscapeHeightmapTrackerEdMode : public FEdMode
 {
 public:
 	virtual bool InputKey(FEditorViewportClient* ViewportClient, FViewport* Viewport, FKey Key, EInputEvent Event) override;
+	virtual void Render(const FSceneView* View, FViewport* Viewport, FPrimitiveDrawInterface* PDI) override;
 
 private:
 	bool TraceLandscapeClick(FEditorViewportClient* ViewportClient, FViewport* Viewport) const;